+++ src/main/java/com/qasession/controller/service/QuestionService.java
@@ -1,6 +1,7 @@
 package com.qasession.controller.service;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -9,14 +10,20 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.qasession.controller.dao.AttendeeDao;
 import com.qasession.controller.dao.QuestionDao;
 import com.qasession.controller.dao.SessionDao;
+import com.qasession.controller.dao.UserTranslateDao;
+import com.qasession.controller.model.Attendee;
 import com.qasession.controller.model.Question;
+import com.qasession.controller.security.FacebookClient;
+import com.qasession.controller.security.UserInfo;
 import com.wordnik.swagger.annotations.ApiOperation;
 import com.wordnik.swagger.annotations.ApiResponses;
 import com.wordnik.swagger.annotations.ApiResponse;
@@ -24,31 +31,19 @@ import com.wordnik.swagger.annotations.ApiResponse;
 @Produces("application/json")
 @Path("/session/{sessionId}/question")
 public class QuestionService {
-	private static Logger LOGGER = LoggerFactory.getLogger(QuestionService.class);
-	
-	@Resource(shareable=true, name="getQuestionDao")
+	private static Logger LOGGER = LoggerFactory
+			.getLogger(QuestionService.class);
+
+	@Resource(shareable = true, name = "getUserTranslateDao")
+	private UserTranslateDao mUserTranslateDao;
+
+	@Resource(shareable = true, name = "getAttendeeDao")
+	private AttendeeDao mAttendeeDao;
+
+	@Resource(shareable = true, name = "getQuestionDao")
 	private QuestionDao mQuestionDao;
-	@Resource(shareable=true, name="getSessionDao")
+	@Resource(shareable = true, name = "getSessionDao")
 	private SessionDao mSessionDao;
-	@GET
-	@Path("/")
-	@ApiOperation(value = "Find question by question ID", notes = "Returns all question record that this session belong")
-	@ApiResponses(value = {
-			@ApiResponse(code = 404, message = "Attendee ID not found"),
-			@ApiResponse(code = 403, message = "Not authorized") })
-	public Response getAllQuestion(@PathParam("sessionId") String pSessionId,
-			@PathParam("attendeeId") String pAttendeeId) {
-		try {
-			return Response
-					.ok()
-					.entity(mQuestionDao.getQuestionsByKeyValue("session_id",
-							pSessionId)).build();
-		} // try
-		catch (Exception pExeception) {
-			LOGGER.error(pExeception.toString());
-			return Response.serverError().build();
-		} // catch
-	} // Session getAllAttendeeOccurance
 
 	@GET
 	@Path("/{questionId}")
@@ -59,8 +54,8 @@ public class QuestionService {
 	public Response getQuestionById(@PathParam("sessionId") String pSessionId,
 			@PathParam("questionId") String pQuestionId) {
 		try {
-			return Response
-					.ok()
+
+			return Response.ok()
 					.entity(mQuestionDao.getQuestionById(pQuestionId)).build();
 		} // try
 		catch (Exception pExeception) {
@@ -77,12 +72,32 @@ public class QuestionService {
 			@ApiResponse(code = 403, message = "Not authorized") })
 	public Response updateSessionAttendee(
 			@PathParam("sessionId") String pSessionId,
-			@PathParam("questionId") String pQuestionId, Question pQuestion) {
+			@PathParam("questionId") String pQuestionId, Question pQuestion,
+			@Context HttpServletRequest pHttpServletRequest) {
 		try {
-			pQuestion.setSession(mSessionDao.getSessionById(pSessionId));
-			return Response
-					.ok()
-					.entity(mQuestionDao.createQuestion(pQuestion)).build();
+
+			UserInfo lUserInfo = (UserInfo) pHttpServletRequest.getSession()
+					.getAttribute(FacebookClient.getUserInfoSessionId());
+			Attendee lAttendee = mAttendeeDao.getAttendeeBySessionIdUserId(
+					pSessionId, lUserInfo.getUserId());
+
+			Question lQuestion = mQuestionDao.getQuestionById(pQuestionId);
+
+			if (!lUserInfo.getUserRole().equals("ADMIN")
+					&& (lQuestion.getCreatedBy().getUserTranslate().getUserId()
+							.equals(lUserInfo.getUserId()) == false
+							|| lAttendee == null || !lAttendee.getSessionRole()
+							.equals("HOST") || !lAttendee.getSession().getSessionStatus().equals("OPEN"))) {
+				return Response
+						.status(Response.Status.FORBIDDEN)
+						.entity("{\"status\":\"forbidden\", \"message\":\"user does not have the previliege to perform the particular action\"}")
+						.build();
+			} // if
+
+			pQuestion.setQuestionId(pQuestionId);
+
+			return Response.ok().entity(mQuestionDao.updateQuestion(pQuestion))
+					.build();
 		} // try
 		catch (Exception pExeception) {
 
@@ -99,13 +114,37 @@ public class QuestionService {
 			@ApiResponse(code = 404, message = "Attendee ID not found"),
 			@ApiResponse(code = 403, message = "Not authorized") })
 	public Response createSessionAttendee(
-			@PathParam("sessionId") String pSessionId,
-		    Question pQuestion) {
+			@PathParam("sessionId") String pSessionId, Question pQuestion,
+			@Context HttpServletRequest pHttpServletRequest) {
 		try {
+
+			UserInfo lUserInfo = (UserInfo) pHttpServletRequest.getSession()
+					.getAttribute(FacebookClient.getUserInfoSessionId());
+
+			Attendee lAttendee = mAttendeeDao.getAttendeeBySessionIdUserId(
+					pSessionId, lUserInfo.getUserId());
+
+			if (lAttendee == null || !lAttendee.getSession().getSessionStatus().equals("OPEN")) {
+				return Response
+						.status(Response.Status.FORBIDDEN)
+						.entity("{\"status\":\"forbidden\", \"message\":\"user does not have the previliege to perform the particular action\"}")
+						.build();
+			} // if
+
+			if (mQuestionDao.getQuestionsBySessionIdUserId(pSessionId,
+					lUserInfo.getUserId()).size() > lAttendee.getSession()
+					.getSessionMaxQuestion()) {
+				return Response
+						.status(Response.Status.FORBIDDEN)
+						.entity("{\"status\":\"forbidden\", \"message\":\"user exceeds allowed questions in the session\"}")
+						.build();
+			} // if
+
+			pQuestion.setCreatedBy(lAttendee);
 			pQuestion.setSession(mSessionDao.getSessionById(pSessionId));
-			return Response
-					.ok()
-					.entity(mQuestionDao.createQuestion(pQuestion)).build();
+
+			return Response.ok().entity(mQuestionDao.createQuestion(pQuestion))
+					.build();
 		} // try
 		catch (Exception pExeception) {
 			return Response.serverError().build();
@@ -120,12 +159,32 @@ public class QuestionService {
 			@ApiResponse(code = 403, message = "Not authorized") })
 	public Response deleteSessionAttendee(
 			@PathParam("sessionId") String pSessionId,
-			@PathParam("questionId") String pQuestionId) {
+			@PathParam("questionId") String pQuestionId,
+			@Context HttpServletRequest pHttpServletRequest) {
 		try {
+
+			UserInfo lUserInfo = (UserInfo) pHttpServletRequest.getSession()
+					.getAttribute(FacebookClient.getUserInfoSessionId());
+			Attendee lAttendee = mAttendeeDao.getAttendeeBySessionIdUserId(
+					pSessionId, lUserInfo.getUserId());
+
+			Question lQuestion = mQuestionDao.getQuestionById(pQuestionId);
+
+			if (!lUserInfo.getUserRole().equals("ADMIN")
+					&& (lQuestion.getCreatedBy().getUserTranslate().getUserId()
+							.equals(lUserInfo.getUserId()) == false
+							|| lAttendee == null || !lAttendee.getSessionRole()
+							.equals("HOST") || !lAttendee.getSession().getSessionStatus().equals("OPEN"))) {
+				return Response
+						.status(Response.Status.FORBIDDEN)
+						.entity("{\"status\":\"forbidden\", \"message\":\"user does not have the previliege to perform the particular action\"}")
+						.build();
+
+			} // if
+
 			mQuestionDao.deleteQuestionById(pQuestionId);
-			return Response
-					.ok()
-					.entity("{\"status\":\"success\"}").build();
+
+			return Response.ok().entity("{\"status\":\"success\"}").build();
 		} // try
 		catch (Exception pExeception) {
